# leaseservice/src/main/resources/swagger/car.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8091
    description: Generated server url
paths:
  /cars/{id}:
    get:
      tags:
        - car-controller
      operationId: getCarById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Car'
    put:
      tags:
        - car-controller
      operationId: updateCar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Car'
    delete:
      tags:
        - car-controller
      operationId: deleteCarById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
  /cars/leaserate:
    post:
      tags:
        - car-controller
      operationId: getLeaseRate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaseRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LeaseRateResponse'
  /cars/:
    get:
      tags:
        - car-controller
      operationId: getAllCars
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Car'
    post:
      tags:
        - car-controller
      operationId: createCar
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Car'
components:
  schemas:
    CarRequest:
      type: object
      properties:
        make:
          type: string
        model:
          type: string
        version:
          type: string
        co2Emission:
          type: string
        noOfDoors:
          type: integer
          format: int32
        grossPrice:
          type: number
        netPrice:
          type: number
    Car:
      type: object
      properties:
        id:
          type: integer
          format: int64
        make:
          type: string
        model:
          type: string
        version:
          type: string
        co2Emission:
          type: string
        noOfDoors:
          type: integer
          format: int32
        grossPrice:
          type: number
        netPrice:
          type: number
    LeaseRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        mileage:
          type: integer
          format: int32
        durationInMonths:
          type: integer
          format: int32
        interestRate:
          type: number
    LeaseData:
      type: object
      properties:
        mileage:
          type: integer
          format: int32
        durationInMonths:
          type: integer
          format: int32
        interestRate:
          type: number
        car:
          $ref: '#/components/schemas/Car'
    LeaseRateResponse:
      type: object
      properties:
        leaseData:
          $ref: '#/components/schemas/LeaseData'
        leaseRate:
          type: number